Make markup validate against https://validator.w3.org
These changes mean that the HTML generates no errors or warnings when
ran against the W3C's HTML5 validator.

Before we had <h> tags inside the main <ol> which is not valid. This
commit
  - replace the <li> tags that wrap each paragraph with a <div>
  - removes the top level <ol>
  - adds semantic <p> tags around the paragraphs in each div

In doing so we lose the default numbering which we got for free by
basing everything around a single <ol>. We address that using CSS in a
subsequent commit.